added GetVentData

+++ Models/DashboardData.cs
@@ -2,7 +2,7 @@ namespace smart_vents_api.Models;
 
 public class DashboardData
 {
-    public float TargetTemp { get; set; }
+    public double TargetTemp { get; set; }
 
     public bool IsEnabled { get; set; }
 }
